requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-apscheduler==0.6.2
django-compat==1.0.15
django4-background-tasks==1.2.7
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
whitenoise==6.5.0
